Using vim-pathogen as VIM package manager
With vim-pathogen, different vim packages can reside in their own
directory. This in turn allows to easily maintain and upgrade the
different VIM packages by directly keep them as git submodules under
the bundle/ directory.

Keep in mind this is allowed but *not* endorsed by Tim Pope, the
current maintainer of vim-pathogen.

The vimtodo package has been added, either because useful and to test
vim-pathogen.