Changed xoria256 with vim-wombat
The xoria256 theme piddles with the TERM variable and barks every time
it is not set (or if is set to a non 256color value). I use a lot of
different PCs (locally and throught ssh), mainly ubuntu, debian and
archlinux, and noone has the TERM variable set: I'm freaking tired of
setting TERM variables.

The (new) wombat256 does not have this check and seems to not suffer
from the drawback of the original wombat theme (different scheme in text
terminal), so I'm giving it a try.